controllers: fix updating of ingress, service and secrets
This commit fixes the update of the owned ingress, service and secret resources
by using `Client.Update' instead of `Client.Patch'.

Previous the code was patching the owned resources by creating a baseline, but
without actually modifying/mutating the objects themselves. This resulted in a
no-op patch operation, which in the end didn't update the objects.

Since we already know the final desired ingress, service and secret objects it
makes sense to simply update the resources.